+++ CHANGES.rst
@@ -16,6 +16,7 @@ in progress
 - Add operating system support for RHEL UBI's "minimal" flavors
 - Add custom exception objects
 - Add more test cases to raise code coverage to >75%
+- Refactoring: Bundle ``systemd-nspawn``/``machinectl`` to ``postroj.backend.nspawn``
 
 
 2022-05-08 0.1.0